Refactor to use ClassFields and ES6 arrow functions.
The commit's purpose is to reduce the complexity from the `.bind(this)` and remove constructors.

As ES6 arrow functions don't need to be bound to the context of the class (as it is lexically scoped), this removes the need for `.bind(this)`.

Then state can be moved up as a class field, thus also reducing code by removing the constructor. More code removed, the better :P